Looks like a working version rewritten to use threads.
