+++ dynamo_consistency/_version.py
@@ -2,4 +2,4 @@
 Hold the version here
 """
 
-__version__ = '2.1.6'
+__version__ = '2.2.0'
