+++ dynamo_consistency/remotelister.py
@@ -10,7 +10,7 @@ import logging
 
 from . import config
 from . import datatypes
-from . import threads
+from . import create
 
 from .cache import cache_tree
 from .backend import get_listers
@@ -26,7 +26,7 @@ def listing(site, callback=None):
 
     :param str site: The site name
     :param function callback: The callback function to pass to
-                              :py:func:`threads.create_dirinfo`
+                              :py:func:`create.create_dirinfo`
     :returns: The site directory listing information
     :rtype: dynamo_consistency.datatypes.DirectoryInfo
     """
@@ -36,7 +36,7 @@ def listing(site, callback=None):
     config_dict = config.config_dict()
 
     directories = [
-        threads.create_dirinfo(
+        create.create_dirinfo(
             config_dict['RootPath'], directory, constructor, params, callback)
         for directory in config_dict.get('DirectoryList', [])
         ]
